requirements.txt
google-genai
python-dotenv
google-genai[aiohttp]
pillow
fastapi[standard]